Handle empty multi-party dialogs.
Some sloppy removal of SDH elements could leave the original dash
intact, thus making it look like it is a multi-party dialog but there's
now only a single speaker.

In this case, dialogs are searched for but if it is empty, it will be
removed in the process and continue searching. Without the multiparty
flag set, it will be (correctly) treated as a single dialog and the
only dash would be removed.