Reworked SDH detection.
Previously a single pattern was used to detect the presence of an SDH
element. However, there were false positives if there is dialog
sandwiched between two SDH elements, like "[a] foo [b]".

The algorithm has been reworked to treat the pattern as a unit, remove
all matches of it, then see what remains. If it is not an empty string,
it means there was dialog in there. Additionally, if only dashes and
whitespace remain, it was probably a multi-party dialog but only with
SDH elements, so that is to be removed as well.

Also, because Lua can't handle UTF-8 strings, the music note is now
treated as a single string pattern to be matched.