Handle dialogs embedded within a style
Reworked preceding_newline() to skip style tags and continue searching,
and return the position of the found newline or the string start. This
would handle situations when the dialog dash (or entire line) was
enclosed within a style. Also, when stripping the dash, preserve
contents before the dash, in this case it would be the start style tag.