Reworked loop to handle all dialogs
Previously there was an initial pattern search and the subsequent
dialogs would be handled using the loop. However, the loop also does
things like strip away empty dialogs, which could happen on the first
dialog.

The loop now handles all dialogs, and if there was only one dialog, then
the pattern has to be searched for again because modifications could
have been made during the loops, possibly even to the first dialog.