Bugfix to handle trailing empty dialogs
If a dialog happens not on the first line (char), it will not be
recognized as dialog at all; this fix handles that.

Additional logic was already in place to make sure that if single
dialogs happen NOT on the first line, they will not be removed, even
though technically they qualify as a "single dialog".