Remove SDH elements in dialog also
Currently the script works by removing all known SDH elements, then
stripping newlines and formatting, then dialog dashes and seeing what's
left. If nothing remains, then the entire line is deleted, otherwise it
remains unchanged.

The problem with this approach is that SDH elements in a multi-party
dialog will not be touched, because the other party's dialog is not SDH
and thus not "removable". To approach this, the utility functions from
the fix-single-dialog script was abstracted out so that we can use it
here. With these utilities, we can easily target a single line and able
to remove that line only, in the case of an SDH element. A fix-up pass
using fix-single-dialog should then be applied to remove what remains.