Styles that straddle dialogs may be left untouched
Sometimes it happens, styles may be incorrectly applied. In this case, a
start tag occurs before the newline, causing the dialog to appear
non-empty. Note that it is hard to generically fix these styling
errors, because if you remove the style tag as part of the dialog, what
happens its corresponding style tag? Would that be dangling now?

This commit aims to fix the easiest case, which is when the start style
tag occurs immediately before a newline. The tag is simply moved after,
and the processing code continues normally.