Added ability to remove stylized SDH descriptions in dialog
- Added a skip_tag function to skip past style tags
- Made next_eol and preceding_newline use the skip_tag common implementation

Now when looking for desc_dialog, it tries to use skip_tag to extend the
start of the matching dialog past any style tags, then see if the dialog
and tags are self-contained (doesn't affect styles after removal).
If so, it will remove the dialog with the style tags. Otherwise, it will
log that line as problematic and require manual editing later.

Also added tests to check this functionality.