Fix for second and later dialog desc not being removed
After removal of sound descriptions in a dialog, the next search
position was not updated, thereby left pointing further into the string
or even past its end. This bug was only discovered when multiple
descriptions were present but not all removed.

The fix is to point at the removed portion of the string for the
subsequent pass. Also added more tests for this.